# api/openapi.yaml
openapi: 3.1.0
info:
  title: dimsum control API
  version: 1.0.0
  description: |
    All generation and job IDs are decimal strings. Content revisions and transaction IDs
    are opaque SHA256 strings. No unconfigured statistics dependency returns fabricated data.
    Browser-session mutations require Origin and X-CSRF-Token obtained at login.
    Opaque bearer tokens grant full administrator access without browser Origin/CSRF
    credentials. Tokens have a separate lifetime from browser sessions and remain
    valid until explicitly revoked. At most 32 tokens may be active; plaintext is
    returned only at creation. Collections use YAML
    schema field names. PATCH paths are relative to the collection (including index),
    or absolute configuration paths for settings. Unsupported comment-preserving edits
    fail explicitly. Successful saves include saved and active configuration identities.
    History windows are half-open [from,to), positive, and at most 366 days. Timestamps
    use UTC with at most microsecond precision. Defaults end at the current UTC minute
    and start one hour earlier; when only to is provided, from defaults to to minus
    one hour. All endpoints share these defaults. updated_at is the response computation
    time, not a promise of durable observation coverage; complete carries that meaning.
    Reuse explicit range values across requests for a fixed dashboard window.
...
security:
  - session: []
  - bearer: []
paths:
...
  /api/v1/openapi.json:
    get:
      summary: Read the authenticated OpenAPI document as JSON
      responses:
        '200':
          description: Authoritative OpenAPI document
          content:
            application/json:
              schema: {type: object}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/openapi.yaml:
    get:
      summary: Read the authenticated OpenAPI document as YAML
      responses:
        '200':
          description: Authoritative OpenAPI document
          content:
            application/yaml:
              schema: {type: string}
        default: {$ref: '#/components/responses/Error'}
...
  /session:
    post:
      security: []
      summary: Authenticate the single administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: {type: string}
      responses:
        '200':
          description: HttpOnly SameSite=Strict cookie and CSRF token
          content:
            application/json:
              schema:
                type: object
                required: [csrf_token, expires_at]
                properties:
                  csrf_token: {type: string}
                  expires_at: {type: string, format: date-time}
        default: {$ref: '#/components/responses/Error'}
    delete:
      summary: Revoke current session
      security:
        - session: []
      responses:
        '200': {description: Logged out}
        default: {$ref: '#/components/responses/Error'}
...
  /api/v1/jobs:
    x-mcp-tools:
      get: {name: list_jobs, description: Inspect the last 32 process-lifetime jobs and their results}
      post: {name: create_job, description: Start a refresh or backup or restore or diagnostic job and inspect its result with list_jobs}
    get:
      responses:
        '200':
          description: Last 32 jobs, process lifetime only
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/Job'}}
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind]
              properties:
                kind: {type: string, enum: [refresh, backup, restore, upstream-probe, support-bundle, dhcp-check]}
                input:
                  description: 'For upstream-probe use UpstreamProbeInput; for support-bundle use SupportBundleInput. Diagnostic input is capped at 4 KiB. Validation failures are reported in the asynchronous job status.'
              allOf:
                - if:
                    properties:
                      kind: {const: dhcp-check}
                  then:
                    properties:
                      input: {$ref: '#/components/schemas/DHCPCheckInput'}
                - if:
                    properties:
                      kind: {const: upstream-probe}
                  then:
                    required: [input]
                    properties:
                      input: {$ref: '#/components/schemas/UpstreamProbeInput'}
                - if:
                    properties:
                      kind: {const: support-bundle}
                  then:
                    properties:
                      input: {$ref: '#/components/schemas/SupportBundleInput'}
      responses:
        '202':
          description: Accepted; one running job maximum
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Job'}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/events:
    get:
      responses:
        '200':
          description: 'SSE: reset requires summary refetch; status includes configuration and jobs, at most once/second. Eight clients, two-second write deadline. No replay history.'
          content:
            text/event-stream:
              schema: {type: string}
        default: {$ref: '#/components/responses/Error'}
components:
...
  securitySchemes:
    session: {type: apiKey, in: cookie, name: dimsum_session, description: Browser-session authentication; mutations also require Origin and X-CSRF-Token}
    bearer: {type: http, scheme: bearer, description: Opaque full-administrator token with independent lifetime until explicit revocation; checked on every HTTP request including MCP requests; no browser Origin/CSRF credentials required}
  schemas:
...
    UpstreamProbeInput:
      type: object
      additionalProperties: false
      required: [endpoint]
      properties:
        endpoint: {type: string, description: Literal IP:port currently configured in the active primary or fallback pool; arbitrary destinations are rejected}
        timeout_ms: {type: integer, minimum: 50, maximum: 3000, default: 1000}
...
    SupportBundleInput:
      type: object
      additionalProperties: false
      description: History is omitted by default. Explicit opt-in also requires an explicit positive UTC from/to range of at most 24 hours. Range and limit are rejected without opt-in.
      properties:
        include_query_history: {type: boolean, default: false}
        from: {type: string, format: date-time}
        to: {type: string, format: date-time}
        limit: {type: integer, minimum: 1, maximum: 200, default: 50}
...
    DHCPCheckInput:
      type: object
      additionalProperties: false
      description: Explicit job checks configured interface/static address and socket permissions plus bound DNS listener and configuration conflicts; it does not scan every pool address or prove LAN/firewall reachability. At most one job runs. Input at most 4 KiB. Inspect results via list_jobs.
      properties:
        probe_other_servers: {type: boolean, default: false, description: Explicitly send one DISCOVER on the configured interface; never REQUEST a lease. Observe at most 64 packets and 16 other server IDs. No offer is not proof of absence. Requires Linux and exclusive UDP 68.}
        timeout_ms: {type: integer, minimum: 100, maximum: 3000, default: 1000, description: Other-server observation deadline in milliseconds}
...
    Job:
      type: object
      required: [id, kind, state, created]
      properties:
        id: {type: string, pattern: '^[0-9]+$'}
        kind: {type: string}
        state: {type: string, enum: [running, succeeded, failed]}
        created: {type: string, format: date-time}
        result:
          description: 'Kind-specific result. upstream-probe returns UpstreamProbeResult; support-bundle returns SupportBundleResult; dhcp-check returns DHCPCheckResult. Other job kinds retain their documented result shapes.'
        error: {type: string}
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id, field_errors]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string}
            active_generation: {type: string, pattern: '^[0-9]+$'}
            field_errors: {type: array, items: {type: object}}
  responses:
...
    Error:
      description: '400 malformed/bad_request, 401 auth, 403 CSRF/Host/Origin, 404 missing or expired detail, 409 revision/generation conflict, 422 validation, 429 capacity, 503 unavailable'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}